Add comments explaining AppProcessFixture close

+++ src/test/java/com/eighthlight/fabrial/test/http/AppProcessFixture.java
@@ -38,10 +38,20 @@ public class AppProcessFixture implements AutoCloseable {
   public void close() {
     appProcess.destroy();
     var result = Result.attempt(() -> appProcess.waitFor(10, TimeUnit.SECONDS));
+
+    // re-interrupt if interrupted
     result.getError().ifPresent(inte -> Thread.currentThread().interrupt());
+
+    // capture exit value from first "destroy" attempt (or prior crash)
     var exited = result.getValue().get();
     var exitValue = appProcess.exitValue();
-    appProcess.destroyForcibly();
+
+    // if initial destroy didn't work, try harder
+    if (!exited) {
+      appProcess.destroyForcibly();
+    }
+
+    // fail the test if the app didn't exit cleanly
     assertThat(exited, is(true));
     assertThat(exitValue, is(143));
   }
